+++ tools/RAiDER/checkArgs.py
@@ -35,6 +35,12 @@ def checkArgs(args):
     #########################################################################################################################
     # Date and Time parsing
     args.date_list = [datetime.combine(d, args.time) for d in args.date_list]
+    if (len(args.date_list) > 1) & (args.orbit_file is not None):
+        logger.warning('Only one orbit file is being used to get the '
+                        'look vectors for all requested times, if you '
+                        'want to use separate orbit files you will '
+                        'need to run raider separately for each time.')
+    args.los.setTime(args.date_list[0])
 
     #########################################################################################################################
     # filenames